requirement.txt
openai 
langchain
python-dotenv
streamlit
PyPDF2

-e . 